scripts/requirements.txt
feedparser
requests
python-dotenv
requests-oauthlib
